requirements.txt

# requirements.txt
Flask==2.3.3
Flask-Session==0.5.0
numpy==1.24.3
pandas==1.5.3
scikit-learn==1.3.0
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.2
